+++ src/01_Foundations/01_Brand.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Iconography" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Brand" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Iconography
+# Brand Guidelines
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
