+++ src/01_Foundations/02_Accessibility.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Brand" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Accessibility" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Brand
+# Accessibility Standards
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
