+++ src/01_Foundations/03_Typography.stories.mdx
@@ -1,6 +1,6 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Typography" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Typography" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
 # Typography Guidelines
