+++ src/01_Foundations/04_Naming.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Illustrations" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Naming" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Illustrations
+# Naming Conventions
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
 
@@ -16,6 +16,7 @@ It can, however, still contain a `<Canvas>` definition:
     <p>
       You may need to wrap certain documentation in a `Preview` in order to reset the styles for use
       cases such as documenting a design system.
+      Case Styles: Camel, Pascal, Snake, and Kebab Case.
     </p>
   </div>
 </Canvas>
