+++ src/01_Foundations/05_Grid.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Grid" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Grid" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Grid
+# Grid Layout
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
