+++ src/01_Foundations/06_Illustration.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Naming Conventions" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Illustration" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Naming Conventions
+# Illustration Library
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
 
@@ -16,7 +16,6 @@ It can, however, still contain a `<Canvas>` definition:
     <p>
       You may need to wrap certain documentation in a `Preview` in order to reset the styles for use
       cases such as documenting a design system.
-      Case Styles: Camel, Pascal, Snake, and Kebab Case.
     </p>
   </div>
 </Canvas>
