+++ src/01_Foundations/07_Icons.stories.mdx
@@ -1,8 +1,8 @@
 import { Meta, Canvas } from '@storybook/addon-docs';
 
-<Meta title="Guidelines/Accessibility" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
+<Meta title="Foundations/Icons" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Accessibility
+# Iconography
 
 This file is a documentation-only MDX file, i.e. it doesn't contain any `<Story>` definitions.
