+++ src/02_DesignTokens/023_Icons/03_Icons.stories.mdx
@@ -4,6 +4,6 @@ import IconSizes from '../023_Icons/01_IconSizes/IconSizes';
 
 <Meta title="Design Tokens/Icons" parameters={{ previewTabs: { canvas: { hidden: true } } }} />
 
-# Iconography Design Tokens
+# Icon Design Tokens
 
 <IconSizes />
