+++ src/10_Styles/Main.scss
@@ -21,7 +21,7 @@
 @import "../10_styles/05_Typography";
 
 // Atom Styles
-@import "../03_Atoms/Iconography/Icon";
+@import "../03_Atoms/Icon/Icon";
 @import "../03_Atoms/Button/Button";
 @import "../03_Atoms/ExLink/ExLink";
 @import "../03_Atoms/Figure/Figure";
