+++ src/10_Styles/_03_DesignTokens.scss
@@ -40,9 +40,9 @@
 // Tracking Design Tokens
 @import "../02_DesignTokens/022_Typography/05_Tracking/Tracking";
 
-// 003 Iconography Design Tokens
+// 003 Icons Design Tokens
 /*******************************/
-// Icon Design Tokens
+// Icon Size Design Tokens
 @import "../02_DesignTokens/023_Icons/01_IconSizes/IconSizes";
 
 // 004 Layout Design Tokens
