+++ src/pages/style-guide.js
@@ -6,7 +6,7 @@
 import React from "react"
 import Layout from "../06_Templates/MainLayout/MainLayout"
 import Figure from "../03_Atoms/Figure/Figure"
-import Icon from "../03_Atoms/Iconography/Icon"
+import Icon from "../03_Atoms/Icon/Icon"
 import InLink from "../03_Atoms/InLink/InLink"
 import ExLink from "../03_Atoms/ExLink/ExLink"
 import Button from "../03_Atoms/Button/Button"
